machine-learning/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
ollama==0.4.7
langchain==0.3.17
langchain-community==0.3.16
faiss-cpu==1.10.0